Updates for fltk-1.4 compatibility
For convenience the path to the fltk-config can be specified to the
Makefile, which allows to build against different fltk versions easily.

Fixed appicon00 project to compile with fltk-1.4 as well

The touchbar00 project’s makefile is also updated, but not other changes
were needed, it already compiled with fltk-1.4

The touchbar01 project is updated also due to changes in the platform
driver.